business_seconds/pyproject.toml
[project]
name = "business-seconds"
version = "0.1.0"
description = "A Django API to calculate business seconds between two times"
authors = [{name = "Your Name", email = "your.email@example.com"}]
dependencies = [
    "django>=4.2.0",
    "djangorestframework>=3.14.0",
    "python-dateutil>=2.8.2",
    "requests>=2.31.0",
    "psycopg2-binary>=2.9.0",
    "drf-spectacular>=0.26.0",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=5.0.0",
    "coverage>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["calculator", "business_seconds_api"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.coverage.run]
source = ["."]
omit = [
    "*/migrations/*",
    "*/venv/*",
    "*/env/*",
    "manage.py",
    "*/settings/*",
    "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
